Remove need for setup_spice
